packable: only pass visitor to verify_with if it was provided (#72)
* packable: only pass visitor to verify_with if it was provided

* Some rename

* Prepare release

* Fix tests

* provided -> explicit

* filter_attrs

* Bump dates

* Cleanup

* Add breaks

* Clippy

+++ packable/packable-derive-test/tests/fail/invalid_verify_with_field_type.rs
@@ -21,11 +21,11 @@ impl From<Infallible> for PickyError {
     }
 }
 
-fn verify_value<const VERIFY: bool>(&value: &u8) -> Result<(), PickyError> {
+fn verify_value<const VERIFY: bool>(&value: &u64) -> Result<(), PickyError> {
     if !VERIFY || value == 42 {
         Ok(())
     } else {
-        Err(PickyError(value))
+        Err(PickyError(value as u8))
     }
 }
