+++ packable/packable-derive-test/tests/pass/verify_with_field_visitor.rs
@@ -21,16 +21,17 @@ impl From<Infallible> for PickyError {
     }
 }
 
-fn verify_value<const VERIFY: bool>(&value: &u64, _: &()) -> Result<(), PickyError> {
+fn verify_value<const VERIFY: bool>(&value: &u8, _: &()) -> Result<(), PickyError> {
     if !VERIFY || value == 42 {
         Ok(())
     } else {
-        Err(PickyError(value as u8))
+        Err(PickyError(value))
     }
 }
 
 #[derive(Packable)]
 #[packable(unpack_error = PickyError)]
+#[packable(unpack_visitor = ())]
 pub struct Picky(#[packable(verify_with = verify_value)] u8);
 
 fn main() {}
